int -> float

+++ Game/Get Off Me/Assets/Scripts/Application/Model/SaveGameModel.cs
@@ -14,5 +14,5 @@ public class SaveGameModel {
 
     public int HighestEnemyKillCountWithoutGettingHit;
 
-    public int HighestTimeWithoutLosingHighCombo;
+    public float HighestTimeWithoutLosingHighCombo;
 }
